// src/app/api-services-v2/model/costCenterResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CostCenterResponseDTO { 
    id?: string;
    name?: string;
    begin_date?: string;
    end_date?: string;
    comment?: string;
}