// src/app/api-services-v2/model/personResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PersonResponseDTO { 
    id?: number;
    name?: string;
    surname?: string;
    email?: string;
    fax?: string;
    phone?: string;
    title?: string;
    comment?: string;
    address_id?: number;
    gender?: PersonResponseDTO.GenderEnum;
}
export namespace PersonResponseDTO {
    export const GenderEnum = {
        M: 'm',
        F: 'f'
    } as const;
    export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];
}